Deleted strings directory and moved to allocators directory.
