+++ src/internal/allocators/gstring.c
