+++ src/internal/allocators/gstring.h
