+++ src/internal/allocators/mstring.c
