+++ src/internal/allocators/mstring.h
