+++ src/internal/allocators/script_stack.c
