+++ src/internal/allocators/vstring.c
